% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multisetting_20200519_functions.R
\name{init_para}
\alias{init_para}
\title{initiate default parameters for simulation}
\usage{
init_para(setting = 2, country_id = 1, social_distancing_flg = 1)
}
\arguments{
\item{setting}{integer. rural simulation; 2 for urban, 3 for slum}

\item{country_id}{1 stands for Uganda demographic}

\item{social_distancing_flg}{1 stands for no social distancing}
}
\value{
a list contains all parameters at default setting for each scenatrio
}
\description{
initiate default parameters for simulation
}
\examples{
# initialise a normal para <- init_para()
}
