% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multisetting_20200519_functions.R
\name{network_generate}
\alias{network_generate}
\title{Simulate synthetic population estimate ERGM contact networks.}
\usage{
network_generate(para, output = "example", searched_clustering_number = 4)
}
\arguments{
\item{para}{list contains all parameter, which should be from the function init_para; the paramter settings could be modified at this stage}

\item{output}{character string for the function to save the output into. The function will create and save results into "Networks" directory.}

\item{searched_clustering_number}{a parameter determin the strength of clustering within the population. Should use default setting unless
for very specific testing session (e.g. simulating lockdown scenario where there is no contact except designated shopping time)}
}
\value{
a list of two object: first is an ERGM object which could be used to simulate contact networks; second is a para object which was updated
with the synthetic population age and family information.
}
\description{
Simulate synthetic population estimate ERGM contact networks.
}
\examples{
para <- init_para()
nw <- network_generate(para)
plot(simulate(nw[[1]]))
}
