% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multisetting_20200519_functions.R
\name{simulate_transmission}
\alias{simulate_transmission}
\title{Title}
\usage{
simulate_transmission(
  NW_SIM = NA,
  input_location = "Networks/example.network.Rdata",
  PCR = F,
  RDT = F,
  len_sim = 100
)
}
\arguments{
\item{NW_SIM}{A list of two object, which is the output of network_generate function.}

\item{input_location}{A file location to read the NW_SIM object.}

\item{PCR}{logical True means PCR test is deployed for the containment}

\item{RDT}{logical True means rapid diagnostic tests are deployed for the containment}

\item{len_sim}{numeric number of days to simulate for the outbreak, default 100.}
}
\value{
a dataframe contains the simulated results.
}
\description{
Title
}
