% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_epidemic_curves}
\alias{plot_epidemic_curves}
\title{Title}
\usage{
plot_epidemic_curves(results, title_fig = "")
}
\arguments{
\item{results}{Output of the network_covid_simulate function}

\item{title_fig}{Give a title of the figure}
}
\value{
A ggplot figure object which plot the epidemic curves under four conditions.
}
\description{
Title
}
\examples{
 results <- network_covid_simulate(rep_num = 1, network_num = 1, output = "example", para = NA)
plt <- plot_epidemic_curves(results, title_fig = "")

}
